feat(routes): implement route search and microbus tracking logic
- Implemented repositories methods for counting microbuses on the way and in station queues
- Integrated UnitOfWork with RouteService
- Added AutoMapper mappings for route and microbus responses

+++ SmartMicrobus.Core/DTO/Route/RouteSummaryResponse.cs
@@ -3,7 +3,7 @@
     public class RouteSummaryResponse
     {
         public decimal Price { get; set; }
-        public decimal DistanceKm { get; set; }
+        public double DistanceKm { get; set; }
         public int NumberOfMicrobusesInQueue { get; set; }
         public int NumberOfMicrobusesOnTheWay { get; set; }
         public int NearestArrivalMinutes { get; set; }
